+++ nodes/__init__.py
@@ -2,5 +2,6 @@
 
 from .Camera import Camera
 from .Controller import Controller
+from .Sensor import Sensor
 
-__all__ = ['Camera', 'Controller']
+__all__ = ['Camera', 'Controller', 'Sensor']